requirements.txt
langchain
openai
chromadb
streamlit
tiktoken
python-dotenv
langchain-community
# Additional libraries for specific functionalities

